Refactor home and marketing layout composition
